requirements-minimal.txt
# Minimal dependencies for Paper Trading Platform
# Ultra-lightweight version

# Data processing (essential)
pandas>=2.0.0
numpy>=1.24.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# System monitoring
psutil>=5.9.0

# Async utilities
aiofiles>=23.0.0

# HTTP client
requests>=2.28.0

# Trading API
KiteConnect>=4.0.0

# YAML support
PyYAML>=6.0

# Async support
asyncio-mqtt>=0.16.0
